refactor python examples and add general type checking and autoformatting
