+++ python/pyproject.toml
@@ -5,6 +5,12 @@ description = ""
 authors = ["Your Name <you@example.com>"]
 readme = "README.md"
 
+[tool.poe.tasks]
+lint = ["pyright", "format-check"]
+pyright = "pyright"
+format-check = "black --diff --check --exclude aptos ."
+format = "black --exclude aptos ."
+
 [tool.poetry.dependencies]
 python = "^3.11"
 grpcio = "^1.53.0"
@@ -18,7 +24,12 @@ pydantic = "^1.10.7"
 [tool.poetry.group.dev.dependencies]
 grpcio-tools = "^1.53.0"
 black = "^23.3.0"
+pyright = "^1.1.305"
+poethepoet = "^0.19.0"
 
 [build-system]
 requires = ["poetry-core>=1.4.2"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.pyright]
+exclude = ["aptos"]
